feat(sdk): complete the two primitives — Commands(execute+provide) + Events(subscribe+emit)
Split into the CLAUDE.md two-primitives shape over a shared facade Transport:
- transport.ts — the facade binding: FOUR methods (execute, provide, subscribe,
  emit) = two primitives × two directions. Shared types + buildCommandUri + Target.
- Commands.ts — Command primitive: execute<K> (name-inferred, cross-env target)
  + provide<K> (serve/client-provided).
- Events.ts — Event primitive: subscribe<K> (payload-inferred) + emit<K> (publish,
  rides AircEventPublisher cross-grid).

Surfaces the remaining facade gaps: #1663 shipped execute+subscribe; provide +
emit are the serve/publish sides to bind in the same uniffi .udl pass so the
native binding is complete in one shot. Zero logic in the SDK; CommandMap/EventMap
generated.
